refactor(suite): fix circulars in component props
