+++ packages/suite/src/components/suite/graph/TransactionsGraph/GraphTooltipDashboard.tsx
@@ -3,10 +3,10 @@ import { type TooltipProps } from 'recharts';
 import { useFormatters } from '@suite-common/formatters';
 import { BASE_CURRENCY_ZERO } from '@suite-common/wallet-utils';
 
+import type { FiatGraphProps } from 'src/components/suite/graph/types';
 import { type CommonAggregatedHistory, type GraphRange } from 'src/types/wallet/graph';
 
 import { GraphTooltipBase } from './GraphTooltipBase';
-import type { FiatGraphProps } from './TransactionsGraph';
 
 interface GraphTooltipDashboardProps extends TooltipProps<number, any> {
     selectedRange: GraphRange;
